backend/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.3
django-celery-results==2.5.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
kombu==5.3.5
prompt-toolkit==3.0.43
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
# uWSGI==2.0.24
vine==5.1.0
wcwidth==0.2.13
flower==2.0.1
psutil==5.9.0
django-celery-beat==2.6.0